+++ src/index.ts
@@ -111,6 +111,7 @@ export type {
   GanttConstraint,
   GanttConstraintType,
   GanttContext,
+  GanttCreateEvent,
   GanttDependencyChange,
   GanttDependencyEvent,
   GanttDependencyUpdate,